+++ src/main/kotlin/com/blackfireweb/stryker/icons.kt
@@ -4,5 +4,5 @@ import com.intellij.openapi.util.IconLoader
 import javax.swing.Icon
 
 object PluginIcons {
-    val STRYKER: Icon = IconLoader.getIcon("/icons/stryker-16x16.png")
+    val STRYKER: Icon = IconLoader.getIcon("/icons/stryker-16x16.png", PluginIcons.javaClass)
 }